Project done in package tracker
This project has been one of my favorites because it entails many hours of effort, protocols such as OSPF were used for both ipv4 and ipv6 networks, access to the cloud was used, a TFTP in case of disasters, web servers with their respective HTTP, has ACL assigned so that certain offices in different countries cannot enter the web. In short, it was a very extensive work that you could download if you want to see it more thoroughly.